Remove unnecessary generic `Callback` class
