+++ project/algorithms/callbacks/__init__.py
@@ -1,9 +1,7 @@
-from .callback import Callback
 from .classification_metrics import ClassificationMetricsCallback
 from .samples_per_second import MeasureSamplesPerSecondCallback
 
 __all__ = [
-    "Callback",
     "ClassificationMetricsCallback",
     "MeasureSamplesPerSecondCallback",
 ]
